src/shop/serializers/category.py
from rest_framework import serializers
from shop.models import Category

class CategoryDetailSerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = '__all__'